requirements.txt
Flask
Flask-SQLAlchemy
flask-smorest
Flask-HTTPAuth
Flask-Migrate

python-dotenv

loguru
